pyproject.toml
[project]
name = "etl-testing-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiofiles==22.1.0",
    "awkward==2.0.7",
    "cocina",
    "colored==1.4.4",
    "emoji==2.2.0",
    "flask==2.2.5",
    "gitpython==3.1.29",
    "hist==2.6.3",
    "ipython>=8.12.3",
    "mplhep==0.3.26",
    "numpy==1.24.2",
    "pandas==1.5.2",
    "pyarrow==11.0.0",
    "pyserial>=3.5",
    "pyqt5>=5.15.11; sys_platform != 'win32'",
    "pyqt5<=5.15.2; sys_platform == 'win32'",
    "pyyaml==5.3.1",
    "requests==2.31.0",
    "rich==13.3.0",
    "scipy==1.8.0",
    "tabulate==0.9.0",
    "tqdm==4.64.1",
    "yahist==1.14.0",
    "pathlib>=1.0.1",
    "etlup>=4.1.2",
]

[tool.uv.sources]
cocina = { git = "https://github.com/danbarto/cocina.git", rev = "v1.0.0" }

[tool.setuptools.packages.find]
where = ["."]